Reformat tor proxy method
'torsocks' was not working with the docker command even though it's supposed to fail if it cannot provide proper masking. This is a redesign of that method using the builtin not well documented '--proxy' flag provided by mansuf/mangadex-downloader original script that I forked into hillmanation/mangadex-downloader-tor to add some functionality to provides the end user confirmation that they are downloading over a proxy. This method not pulls that alternate docker image, as well as pulls the dperson/torproxy docker image and configures it for use on a dedicated container tor network so that the socks5 proxy will work properly in the download containers.